Cleaned up the docker installation

+++ include/arg_parse.h
@@ -16,9 +16,9 @@ namespace speq
             std::filesystem::path io_file_index{"output.idx"};
             std::filesystem::path out_file_path{"output.txt"};
             bool is_force{};
-            u_int threads{std::thread::hardware_concurrency()};
-            u_int chunk{10000};
-            u_int8_t kmer{30};
+            unsigned int threads{std::thread::hardware_concurrency()};
+            unsigned int chunk{10000};
+            unsigned int kmer{30};
             bool is_indexer{false};
             bool is_scanner{false};
             bool is_parsed{false};
